requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
msal
requests
httpx
pandas
beautifulsoup4
lxml
docxtpl
python-docx
pdf2image
pydantic[email]
pydantic-settings>=2.0.0
pytest
black
ruff
slowapi
apscheduler>=3.10.0
# Magic Link Portal dependencies
Pillow>=10.0.0
pypdf>=3.0.0
reportlab>=4.0.0
resend
